Move block_backward_sync from ParallelPlugin to DDPPlugins
